calloc

+++ alloc.h
@@ -1,5 +1,5 @@
 #ifndef _ALLOC
 #define _ALLOC
 #include <stdlib.h>
-void* mylloc(unsigned long long tam);
+void* mylloc(size_t tam);
 #endif
